add pgsql scheme, minor changes

+++ backend/internal/model/credentials.go
@@ -1,6 +1,6 @@
 package model
 
 type Credentials struct {
-	UserID uint64 `gorm:"primaryKey" json:"user_id"`
+	UserID uint   `gorm:"primaryKey" json:"user_id"`
 	Key    string `json:"key"`
 }
